requirements.txt
readabilipy
validators
tqdm
bs4
openai
python-telegram-bot